Move code around to a wavefd package
Split into multiple files. Should work right now. Will populate with the
new classes later.
