+++ fatiando/seismic/wavefd/elastic_sh.py
@@ -1,0 +1,101 @@
+from __future__ import division
+from numpy import sqrt
+import numba
+
+
+def maxdt(area, shape, maxvel):
+    """
+    Calculate the maximum time step that can be used in the simulation.
+
+    Uses the result of the Von Neumann type analysis of Di Bartolo et al.
+    (2012).
+
+    Parameters:
+
+    * area : [xmin, xmax, zmin, zmax]
+        The x, z limits of the simulation area, e.g., the shallowest point is
+        at zmin, the deepest at zmax.
+    * shape : (nz, nx)
+        The number of nodes in the finite difference grid
+    * maxvel : float
+        The maximum velocity in the medium
+
+    Returns:
+
+    * maxdt : float
+        The maximum time step
+
+    """
+    x1, x2, z1, z2 = area
+    nz, nx = shape
+    spacing = min([(x2 - x1) / (nx - 1), (z2 - z1) / (nz - 1)])
+    return 0.606 * spacing / maxvel
+
+
+@numba.jit(nopython=True, nogil=True)
+def nonreflexive_sh_bc(u_tp1, u_t, nx, nz, dt, dx, dz, mu, dens):
+    """
+    Apply nonreflexive boundary contitions to elastic SH waves.
+    """
+    # Left
+    for i in range(nz):
+        for j in range(3):
+            u_tp1[i,j] = u_t[i,j] + dt*sqrt(mu[i,j]/dens[i,j])*(
+                u_t[i,j+1] - u_t[i,j])/dx
+    # Right
+    for i in range(nz):
+        for j in range(nx - 3, nx):
+            u_tp1[i,j] = u_t[i,j] - dt*sqrt(mu[i,j]/dens[i,j])*(
+                u_t[i,j] - u_t[i,j-1])/dx
+    # Bottom
+    for i in range(nz - 3, nz):
+        for j in range(nx):
+            u_tp1[i,j] = u_t[i,j] - dt*sqrt(mu[i,j]/dens[i,j])*(
+                u_t[i,j] - u_t[i-1,j])/dz
+    # Top
+    for j in range(nx):
+        u_tp1[2,j] = u_tp1[3,j]
+        u_tp1[1,j] = u_tp1[2,j]
+        u_tp1[0,j] = u_tp1[1,j]
+
+
+@numba.jit(nopython=True, nogil=True)
+def timestep_sh_esg(u_tp1, u_t, u_tm1, x1, x2, z1, z2, dt, dx, dz, mu, dens):
+    """
+    Perform a single time step in the Finite Difference solution for elastic
+    SH waves using the Equivalent Staggered Grid method.
+    """
+    dt2 = dt**2
+    dx2 = dx**2
+    dz2 = dz**2
+    for i in range(z1, z2):
+        for j in range(x1, x2):
+            u_tp1[i,j] = 2*u_t[i,j] - u_tm1[i,j] + (dt2/dens[i,j])*(
+                (1.125/dz2)*(
+                    0.5*(mu[i+1,j] + mu[i,j])*(
+                        1.125*(u_t[i+1,j] - u_t[i,j])
+                        - (u_t[i+2,j] - u_t[i-1,j])/24.)
+                    - 0.5*(mu[i,j] + mu[i-1,j])*(
+                        1.125*(u_t[i,j] - u_t[i-1,j])
+                        - (u_t[i+1,j] - u_t[i-2,j])/24.))
+                - (1./(24.*dz2))*(
+                    0.5*(mu[i+2,j] + mu[i+1,j])*(
+                        1.125*(u_t[i+2,j] - u_t[i+1,j])
+                        - (u_t[i+3,j] - u_t[i,j])/24.)
+                    - 0.5*(mu[i-1,j] + mu[i-2,j])*(
+                        1.125*(u_t[i-1,j] - u_t[i-2,j])
+                        - (u_t[i,j] - u_t[i-3,j])/24.))
+                + (1.125/dx2)*(
+                    0.5*(mu[i,j+1] + mu[i,j])*(
+                        1.125*(u_t[i,j+1] - u_t[i,j])
+                        - (u_t[i,j+2] - u_t[i,j-1])/24.)
+                    - 0.5*(mu[i,j] + mu[i,j-1])*(
+                        1.125*(u_t[i,j] - u_t[i,j-1])
+                        - (u_t[i,j+1] - u_t[i,j-2])/24.))
+                - (1./(24.*dx2))*(
+                    0.5*(mu[i,j+2] + mu[i,j+1])*(
+                        1.125*(u_t[i,j+2] - u_t[i,j+1])
+                        - (u_t[i,j+3] - u_t[i,j])/24.)
+                    - 0.5*(mu[i,j-1] + mu[i,j-2])*(
+                        1.125*(u_t[i,j-1] - u_t[i,j-2])
+                        - (u_t[i,j] - u_t[i,j-3])/24.)))
